Introduce the language, a little application structure
This commit introduces the language that the application will be built
in: Go. The reason for this is far more centered around me as a
developer rather than for any runtime or practical reasons: I'm most
familiar with it.

The commit also introduces a file called "Development" which is designed
to give casusal observers (read: interviewers looking for an excuse to
look at how I code) an understanding of how I'd prefer to lay things
out.

The commit introduces a subdirectory called "link", which I expect will
contain the link server. If I ever get around to writing this whole
thing, it would likely be:

  — link: the server
  — ui: the user interface
  — api: the api

Or, there might be a binary that combines all functionalities and just
gets run in different "modes". Not sure yet. Let's find out together!

The commit initializes a go module in the root, following the go module
conventions so people can rip libraries from this on GitHub (under the
AGPL license)