Introduce Cobra as the command line interface for the application
There are some more substantial changes in this commit, but the broad
thrust of it is rather than have multiple different binaries, this
commit assumes a single binary that contains all logic that can be
invoked via different "modes" (e.g. server, client and so fourth). The
default mode should probably be client.

== Design Notes
=== Removing Subdirectory "link"

This commit removes the "link" subdirectory, as the original purpose for
this assumed multiple binaries. With the destruction of this, the
subdirectory no longer makes sense and has been removed to simplify the
layout of the application.

=== Cobra Library

This commit introduces the "cobra" library, and a subsequent commit will
probably introduce the "viper" one. These libraries are useful just to
provide an interface to applications, though they are not without some
quirks (e.g. testability of run methods).

The files were generated via:

  $ cobra-cli init

And modified to ensure the license was correct, as well that any
redundant content is removed. Some content was left there to be "filled
out" later, as it is not in version control (and thus cannot be "safely"
deleted")