Introduce clearer error handling for the application
Currently, the application has a single exit code — "os.Exit(1)" given
an error. This is not especially descriptive. There are also a series of
conventions¹ on common ways an application can express why it failed,
without expressing this via a string (sysexits.h)

This commit modifies the application to support these sysexits,
implemented via the previously created andrewhowdencom/sysexits library.

== Design Notes
=== cmd.Execute(c *Cobra.Command) sysexits.Sysexit

This method was created simply to be able to test the code that handles
the coalescence to a sysexit. The broader goal with this work is to have
it as "tested as possible", such that it is clear that I (Andrew) know
how to write testable code.

—
1. https://man.freebsd.org/cgi/man.cgi?query=sysexits&manpath=FreeBSD+4.3-RELEASE