Mock the local deployment of the application
Given that the application is essentially just a link shortener, it will
probably be useful on desktops in the same way the "Golinks"
implementation (or the original Google version) were useful — just as a
pointer to internal resources that are too hard otherwise to discover.

This commit mocks that out. In particular,

== Deployment Documentation

The commit creates deployment documentation, describing what needs to be
done to deploy the application in a given environment. For a local
environment, this is just sticking it somewhere system-wide
(/usr/local/bin/ by convention), and mocking the DNS lookup.

The commit also modifies the README to point to these documents. The
documentation points out setcap, mostly as a demonstration that "yes I
know what capabilities are, and that you do not need to run as root in
order to gain some of the things root can do"

== s3k.link redirect, redirect serve

This commit introduces two new commands that are designed to stub out
the functionality that will exist when the application actually serves
redirects.

The commands are assumed to have a stable API, however the
implementation of the functions is going to change substantially.

=== util

This commit creates a new file called "util", and within that contains
the execute function and a new funciton called "Noop". A "Noop" function
is described, but essentially just means "this command does nothing,
please use a subcommand".

Util is not a good name, but short of another library also called cmd,
i'm not sure what to do with this yet.