Define an adapter for the storage of URLs
The anticipated deployment model of this application assumes a few
different styles of deployment. For example,

  1. Locally, on the host compter
  2. Deployed as a read-only service
  3. Deployed as a read-write service, with writable storage

These things have very different storage constraints. The former can be
served fairly easily by file based configuration, but the latter
requires a more dynamic type of storage.

Given this, this commit introduces a storage "interface". This interface
will be the thing that allows these different storage backends.
Additionally, I anticipate this later is where we will also do the
search / algorithms & data structures work. There are a few ideas, such
as:

  1. Naive, unsorted in-memory storage (that must be O(n))
  2. Naive, sorted in-memory storage (binary search, O(log(n))
  3. Naive, in-memory dict, O(1)
  4. Binary Tree O(n(log(n)))

Which are perhaps to be tested in the coding challenge.