Introduce an example in memory storage
Currently, the application does not have any mechanism to store the
URLs in a way that can be queried by the application (i.e. in memory, or
against an index) to fetch a URL.

This commit addresses that by implementing an in memory storage. It is
perhaps a little more complicated than might be expected, for reasons
that we'll discuss below.

== Design Notes
=== Multiple implementations

As alluded to in other commits, one of the principle outcomes of this
project is demonstrating that I can do the various things expected of a
software engineer — writing tests, writing code, reasoning things
through, commenting and so fourth.

In particular, what I'm concerned about is algorithms and data
structures. While I have a relatively "intuitive" feel for what is
expensive (by reasoning through a graph of its behavior), I struggle to
express this as well. To try and prepare myself for interviews, I'm
implementing a bunch of different efficiencies so that when interviews
ask, I can reason it (in my own memory) against something I've done
recently.

This will be implemented via the in memory implementations of the URL
lookup.

=== Example

This commit includes a "Go self documenting example", just as an example
of how to use the code, as well that I know these functions exist.

=== Benchmark

One of the things that I want to be able to do is encode, via a unit
test (or a benchmark), the difference practically in different
algorithmic implementations. This is makes the whole thing more fun, and
easier for me to reason through rather than the in principle version of
"yes, it makes the following logical steps".

This is expressed by creating a central benchmark in a new "test"
package under storage, and invoking it with different lengths for the
hashmap.

=== Tests

The application is validated in various different ways; that it can
fetch data and retrieve data appropriately. There is also a test
specifically designed to assist the go race detecter to determine
whether the underlying implementation has any race issues.