Shift the concurrency and race detection tests into a central place
The concurrency tests and benchmarks are going to be the same for all
storage implementations. Given this, they can be written once, and run
centrally.

In principle, in future, compliance tests (e.g. "can i read after i
wrote") can also be done centrally, as this will always be the same for
all implementations.