Implement yaml handler
Currently, the storage is only in memory. This is not especially useful
for any use case, save running it for ... temporary redirects? This
commit implements the first useful implementation of this application —
YAML backed storage.

== Design Notes
=== General Design

The Yaml implementation is a thin wrapper around an existing storage
(any storer interface) that loads the content from a Yaml file into that
storage. This allows Yaml to be used with any number of storage engines,
and keeps the implementation very thin.

=== Filesystem

One of the interesting questions to think about was "to abstract the
filesystem, or not to abstract". The initial design was to abstract, but
there's no support for filesystem notifications in these abstractions
(namely, afero).

On balance, I simply ripped the filesystem logic out of the YAML
handler. It is simpler if it reasons against a stream (io.Reader), and
if we need something like a "hot reload" capability we can do this by
replacing the storage engine entirely in the server.

=== The return errors

I went through multiple different designs of this — whether to return
both errors, a single error and so on. Ultimately, for teh API of the
package, I kept the "single typed error" approach
(ErrStorageSetupFailed) as this is all clients should be testing for;
the others are only useful for testing. For testing, I have code
coverage analysis to help me determine whether or not all lines have
been covered.

=== Logger

Normally, I try and avoid using logs as there's often better tools.
However, in this work, there's a class of errors that are ... useful to
know about (corrupted URLs) but also, not something that should
terminate the application themselves. Given this, I'm writing them to
the log output. They're not a trace, nor a metric.

They're debug logs, and can be disabled by overwriting the log
varaible.

=== New testing storage

As part of testing the YAML storage implementation, it was handy to take
a storage that simply returned errors. I expect it will be handy in
future. Given this, it is written.

The API may change substantially — I'm not sure of hte best way to
inject errors just yet.