Connect the storage to the serve command
This commit goes part of the way to implementing the "Serve" command.
Specifically, it implements the command parsing logic and logic to
construct the different storage models in a reasonably elegant, testable
way but the server as quickly and as dirtily as possible. Future work
will rip the server out entirely, after which the command can be
simplified and the server moved to its own package (and subjected to far
more testing).

== Design Notes
=== Testing Cobra

Testing the package for cobra is challenging. Much of it state is
global, and breaking out components for their local counterparts is
challenging as they're not used to being used independently. However,
the existing work seems like a reasonable balance.

=== Factory for storage

One of the interesting questions is "How do we connect the user supplied
configuration with wiring the in". The ideal is probably to
dependency inject the configuration; perhaps type it and make those
types responsible for "sourcing" their own values. However, this is a
lot of work and I haven't got anything taht does this now.

For the moment, the logic is just encoded into this file. This limits
the "leaking" of configuration through the program, and is relatively
testable.

=== Half completed server

The server is half completed, at best. However, with the current state I
can actually use this on my computer in the way described in
DEPLOYMENT.md (as a local link shortener), and I wanted some experience
actually trying it.