Implement absurdly naive (Linear Search) method for finding URLs
This commit implements the absurd case — O(N) (at least, for this type
of data). It just yolo's it all into an array, and scans the array from
the start for each record.

== Design Notes
=== Type of Linear Search

Initially, the type of scan was an array. To populate a new value, the
array pointer needed to be replaced with the pointer to a new array.
This broke the interface (and is probably wildly inefficient memory
allocation), so it is now a struct containing the list.

=== Benchmark Lengths

This commit necessitated having the benchmark lengths be customizable
for the implementations, as the benchmark length of 100k (or 5M) both
took an absurd time to run, and are therefore skipped.

=== Compliance Tests

As the storage implementations do the same, this commit shifts
compliance testing to the central package.