Introduce a binary search implementation
This commit introduces another implementation into the link shortner
designed to demonstrate Olog(n) complexity — Binary Search. This design
takes a data set and recursively cuts it in half until it finds the
value (or not)

For example, to search G

    | A | B | C | D | E | F | G | H |

                  ^ |               |
                          ^ |       |
                            | ^ |

== Design Notes
=== General Implementation

I struggled with this implementation, despite it being conceptually
simple. I struggle with the lexical ordering of content; this is
essentially what the assumption of this work is built on.

For future work, I will likely just assume the array is sorted rather
than demonstrating the sorted insert.