Optimize the README for specific paths
As readers are looking at this project, the anticipated flow is
something like:

1. Get Resume
2. Click on a couple of links
3. Read a little more about a specific task (e.g. code)
4. Decide whether to proceed or no

Alternatively, people may discover this work following an interview,
when they're looking for examples of work that I've previously completed
to either prove, or disprove an understanding they had during the
interview itself.

This commit revises the resume to try and anticipate those tasks, and to
point people to different areas of the code they can discover the
answers to their questions.

== Desgin Notes
=== Missing Areas

There are quite a lot of areas that are missing. Infrastructure as code,
Docker definitions, Kubernetes definitions, Terraform definitions,
Observability, SLOs and so on.

These will all be completed as I actually deploy this application, as a
replacement to h4n.link/pito. To that end, they're only mentioned on the
Resume, where they will be expanded upon as requested.