Implement persistent storage
This commit implements persistent storage via the bbolt package from
etcd. The storage is as simple as it gets — lookup for the value of the
key, insert for the same value.

== Design Notes
=== bbolt.db

BoltDB is just an embedded Go key value store that handles persisting
the content to disk in an efficient way. It seems to be well written,
and widely used — though by what I'm not sure.

=== Setup and Teardown

The test suite for all storage implementations now supports a "setup"
and "teardown" step, and assumes it should exist for all
implementations.

Most of the implementations so far have been in memory, but as we get
more complex here, it seems more likely we'll need to setup and tear
down things.