Split the server into a separate package
This commit modifies the existing implementation of the server, within
the command, with an implementation of the server specific to the
package.

The server in the package is of considerably "higher quality"; that is,
it is better validated with unit tests, written in a way that is more
understandable and easier to extend.

== Design Notes
=== General Pattern

The pattern is now that the server can be create within the command line
and started, as two distinct operations. The servers behavior can be
optionally configured, with the options being present on the command
line.

=== BoltDB

Historically, with the implementation of boltdb, the flags to actually
use it were omitted. This commit amends this in parallel with the server
changes. While BoltDB is not especially useful just yet (as there are no
methods to write URLs to the data store), future work will address this.

=== Gin

Rather than the default HTTP webserver, this commit opts instead to
introduce the gin webserver. Gin takes care of a whole chunk of the
semantic sugar around running a websever, making it easier to define
routes, implement middleware and so on.

Right now, a very limited number of features is implemented — it could
have been done just as well with the standard library. However, future
work — namely, the instrumentation work — will certainly make use of the
middleware implementation.

=== Storage test fix

The storage test implementation had a small bug, in which when
initialized the map storing URLs was nil. This commit addresses it.