Introduce a task management file
Currently, there is no way to understand what the "common tasks" that
might need to be compoleted in this project should be, beyond the
develompent documentation. This commit seeks to address that by
introducing a task runner that will automate the common tasks (at least,
on Linux based platforms).

== Design Notes
=== Why Taskfile

There are many different task runners, from just writing extensive bash
files, using make to using things like Bazel. In my experience, and
especially for Go projects Taskfile provides a suitable interface (via
the Yaml file) with a suitable level of automation (e.g. help,
dependencies, parallellism).

Its also trivial to install and get started for those unfamiliar with
the tooling.

=== Static & Cross Compilation

This commit builds, by default, the binary as a static binary. This is a
common pattern in Go binaries, and is done so that it can later be added
to a "scratch" container without a root filesystem. Its also useful from
a practical point of view, limiting the amount of things that need to be
deployed.

The binary is generated for a combination of operating systems and
architectures. These are the most common ways of deploying this tool I
can reason, and they're going to be later uploaded to GitHub so people
can find them more easily.

=== Help

This commit includes content that is designed to be used with:

  $ task --list
  $ task --summary <command>

Which will help newer contributors understand how to get started with
this project (in parallel with documentation, not yet contributed)