Compress the README
With the work I'm doing in this project, it seems likely that the README
will become pretty busy. This commit reduces the amount of vertical
space occupied by switching from a "heading" style to a "bold" style.

== Design Notes
=== Alphabetical

This commit also shifts the order from a "recency" order to an
"alphabetical" one, to disconnect time from the examples in the project.

=== Loss of anchors

This commit loses anchors to specific titles, but the main entrypoint in
the project is assumed to be Linked In (via h4n.link/otw)