Introduce documentation tooling to the project
This commit introduces tooling to house development and hosting
documentation for the project. In future, it will be deployed to a top
level .dev domain, so that people can use it in their environment.

== Design Notes
=== Empty Documentation

While the documentation tooling is present, there is no documentation.
Future work will add it, following the Divio documentation standard

=== Poetry

The tooling involved is python based, and while it is possible to
install it globally, it is more frequent in python projects to use a
"virtual environment" to house dependencies.

There are various systems to do this, such as pip3 or poetry or
virtualenv, but I have the most recent experience with Poetry and thus
used that.

Future work should reevaluate this choice in favor of something that is
in more frequent use.

=== Mkdocs (& mkdocs-material)

This commit uses the tooling "mkdocs", which is in frequent use in
development projects to generate static websites. It has good support
for this, with client side search.

The mkdocs-material template improves on this with an interface based on
Google's Material UI designs, and additional features built on top of
the standard mkdocs library.