Define the DNS (and infrastructure state artifacts) via Terraform
Currently there is no infrastructure on which to run this application,
host the documentation or do anything else — it is a software
application without application.

This commit starts the process of defining that infrastructure, so that
the application has somewhere practical to run.

== Design Notes
=== Google Cloud

This commit implicitly chooses Google Cloud (specifically, the
"andrewhowdencom" account) to run this application. Google Cloud is just
one of the major cloud providers, and the one I prefer to use. They're
uniquely open with their designs (at least, via academic papers and so)
and their documentation & APIs are all using relatively modern tools.
Tbh, I just like Google technology.

That said, they're much of a muchness.

=== Terraform (well, Tofu)

This commit chooses Terraform (via the now open source Tofu
application) to define its infrastructure. Terraform is a tool that has
wonderful properties like idempotency, a dependency graph and state
management which are all extremely useful when defining infrastructure.
Additionally, it has a large, well supported community and first class
support from the provider.

Alternatives are simply more difficult to run.

=== State via Google Cloud Storage (w. Encryption)

Terraform provides a "state file", which keeps the current state of
infrastructure. This allows us to detect any changes that are made out
of band, and surface them to the implementing developer rather than
simply overwriting any changes as they exist.

The file is, by default, generated and stored on disk. However, it can
include some secret information:

  * https://developer.hashicorp.com/terraform/language/state/sensitive-data

Meaning that the data needs to be kept somewhere secure. Google Cloud
provides the Google Cloud Storage implementation to keep the data
(encrypted):

  * https://cloud.google.com/docs/terraform/resource-management/store-state

This prevents the content being stored in version control and
(inadvertently) leaking data, as well as keeps the data "off disk" in a
way that allows multiple people to work with it.

=== Plan separate to apply

As part of this commit, two tasks are created: "infra-plan" and
"infra-apply". At no time is the infrastructure expected to be deployed
(with changes) without a manual intervention.

The reason for this is practical: Infrastructure exists outside the
"safe change" stream that software itself sits within. Changes in
infrastructure definitions can be destructive and irreversable. Given
this, as a rule, they're expected to change infrequently. Where they do
change, they should be validated by a human.

=== DNS records

This commit defines two new DNS names:

   x40.link → The planned destination for the link shortener
   x40.dev  → The planned destination for the *documentation* for the
              link shortener.