Point x40.dev to GitHub pages
This commit points x40.dev to GitHub pages, which will be the hosting
platform for the documentation for this project.

== Design Notes
=== ALIAS record

Within this work, I discovered that there is a (relatively new?) type of
record called "ALIAS", which functions as a "CNAME" record but for the
apex domain.

=== GitHub Pages

GitHub pages was chosen as while doing work to build the GitHub runner,
I discovered a task which publishes mkdocs based information to GitHub
pages. This is the ideal workflow, and low overhead.