Deploy documentation to GitHub pages
Currently there is (empty) documentation for this project, but
with the ongoing development of features, the goal is to include
documentation within the acceptance criteria for a given feature. This
means that the documentation should be available somewhere for people to
consume.

This commit addresses this problem by deploying the documentation to
GitHub Pages.

== Design Notes
=== GitHub Pages

Originally, the plan was to deploy this documentation into a Google
Cloud Storage bucket. However, in parallel work I discovered the "GitHub
Pages" runners. These runners make it easy enough to deploy static files
to GitHub pages, so they content is instead hosted there.

=== GitHub Actions

Paralell work is exploring a CI/CD implementation for this project. To
this end, I started looking at GitHub actions. Given that they're free,
and they seem to have a fairly sane architecture and good integration
with GitHub, I used them.