Migrate reference documentation to external docs
Currently there is some documentation in the DEVELOPMENT.md file. This
is largely just reference material, to help give people hints that let
them understand the project.

With the publication of the documentation in 8b93609a, this should be
moved out of the project root and into something more discoverable.