Use the "/" character to delimit subtasks
Currently there are different kinds of tasks in the taskfile. For
example, tasks like "clean docs" or similar. These are split with a
"hyphen", which looks visually confusing and can be confused with a
two-word command.

This commit uses the "/" character to delimit commands. This is clearer,
and feels more intuitive, but still also not super satisfying.