Run unit tests in CI
Currently, there is no validation to ensure that as code is changed, the
unit tests will reject any change that breaks compatibility. Rather, the
onus is on the author to run them locally.

This commit seeks to address this by running the unit tests before the
code is merged, preventing the code being "bad" in the mainline branch.
This should prevent accidental breakages.

General Design:
 → Auto merge + deploy. Branch protections. Maybe TF to manage GitHub?

== Design Notes
=== Managing the Git Repository in Terraform

There are a few different ways to communicate the state of the git
repository to other developers; the simplest being to give them access
to the repository itself. However, this being a public project and the
goal to demonstrate skills, another, broader way to demonstrate the
state is to encode it via Terraform.

Given this, this commit elects to start managing the git repository via
terraform, instead of doing it manually. This necessitated in turn
making the deployment a little more complicated, which was subsequently
documented.

=== Merge Queue

The ideal way to do this work is via a "merge queue", which can handle
the work required to promote the build to production (as well as to rip
it out, when that release fails). Unfortunately, this appears to be
enabled only on organizations.

Future work might revisit this once deployed to have such a design, but
for now simply deferring this work users via the "require branch to be
up to date" is sufficient.

=== Test + CGO

Apparently, the tests require cgo enabled. Given this, the commit
overrides the default "disable it".