Lint the project content against common standards (#4)
* Lint the project content against common standards

This commit extends the existing status checks the software validated
against to include "golangci-lints", with the default set of lints.
These allow picking up common issues that can introduce bugs, but that
are not found (so far) in unit tests.

== Design Notes
=== Task vs. Action

The action for GitHub allows pushing the statuses back to the checks
API. Given this, its used instead of using task for the same purpose.

* Quiet the lints

With the enabling of golangci-lint, quite a series of non-checked errors
popped up. While these are not especially interesting (Especially for
tests), it is technically correct to fix them, and thus, they're
addressed through either a small refactoring or validation inline.