Define a container for the application (#5)
While the application currently gets built (and stored in the dist
folder), there is currently not a sufficient package for the application
to be deployed to a production environment.

This commit provides one by packing the application into an OCI
container.

== Design Notes
=== Multi-step builds

The commit uses a multi-step build¹ to first build the binary, and then
copy the binary into a new container with a minimal root filesystem.
This, in principle, allows for a much smaller attack service assuming
_somehow_ an attacker managed to get an RCE within a go application (a
tall order).

In practice, this just allows nice small images.

=== Podman

This commit uses the project "Podman" created by RedHat and donated to
the CNCF as its container tooling. Podman is daemonless, rootless and
allows building and running containers.

It is used instead of Docker as Docker recently imposed some new limits
on the usage of its software, similar to Hashicorp, and I have a
tendency to avoid things unless I understand them.

=== Image Tag

The image tag is the shortref from git. This should be stable, allow a
quick link to Git and shift as the underlying software (or anything
else in the repository) shifts.

=== Imagine Labels

The imagine is enriched with a series of standard "image labels", which
can provide metadata to systems interested in the image itself.

=== Hadolint

Containers are written in a language which is a "psuedo-bash" like
— partially declarative, partially imperative. They inherit the
shortfalls of bash, which can be many.

Given this, this commit includes a linter ("hadolint") for the file.
This linter was chosen simply after scanning a couple of blog posts that
recommend it.

It picked up the distroless image not having a tag, which was
subsequently added (with the sha)

== Links

1. https://docs.docker.com/build/building/multi-stage/
2. https://podman.io/