Build and Store the x40.link container image
This commit introduces a new, mainline pipeline that generates an image
and stores it:

  1. On GitHub, for general public consumption
  2. On Google Cloud, for the (later) deployment to Google Cloud Run

== Design Notes
=== Renaming Pipelines

This commit renames the existing pipelines after a combinatorial of the
event they track, as well as their purpose. This just makes it clearer
to see what should be running whe, just from the file names.

=== Google Cloud Registry

This commit also provides a Google Cloud Registry (via Terraform) in
which to store the images. This is a requirement for Cloud Run (or at
least, makes life much simpler).

This also necessitated the creation of a service account with the
appropriate permissions.

=== Container Changes

While the container build appeared to work locally, there seems to be
small differences between the local environment and production. This
commit makes adjustments to the container (e.g. defining the ARG GOARCH
in all containers) as required to pass the build.

=== GitHub Authentication

Interestingly, Google provides an entirely new mechanism of federated
identities to authenticate GitHub pipelines, rather than doing any sort
of secret exchange.

This is a very ... novel form of doing this, but unless badly
configured, seems orders of magnitude safer. They essentially defer
proving who you are to GitHub, and allow you to define static mappings
for what that user means in Google Cloud.

This is setup via the new "actions.tf" file in the production terraform
deployment folder, and used to push the container to Google Cloud
Artifacts.

=== Ignorable Content

The Google Authentication tooling reportedly adds secrets to the
workspace root, which we want to skip both in Git and in Docker.
However, there's also other things which we can skip in Docker. This
commit skips all of them also.