Add the h4n.me domain (#9)
This work started originally with the goal to replace the h4n.link
domain. That domain is being transferred to another party with the
intent to use it better than I can, so I'll replace it with h4n.me. That
third party has gratiously agreed to 301 all the links from h4n.link to
h4n.me that already exist.

This commit adds h4n.me to the managed domains, such that there can be
nameservers and these can be updated in Gandi. Future work will add
these domains to the link shortener as early tests (via the Yaml
backend, probably).