Limit pipeline execution to changes
Currently, the pipelines are executing every merge to main. This is
producing a lot of artifacts, even though nothing changes. There are a
couple of choices here:

1. Use a specific marker (e.g. a tag) to indicate that a release is
   necessary
2. Run the pipeline conditionally, depending on the changes in the
   request.

This commit takes the latter approach. This should allow a nice "deploy
to master" style approach, but also limit the number of artifacts
generated. It also should clearly indicate which pipelines are skipped
as the tool uses the GitHub API to skip them.

== Design Notes
=== Not using dorny/paths-filter

There's a ... lot of code in this workflow, and I don't know the author.
Given this, I'm choosing a simpler approach for now. See:

* https://github.com/dorny/paths-filter