Use the dorny approach
With more investigation, the dorny plugin appears to be used in Google
artifacts, as well as other large organizations. Additionally, a quick
look indicates there's not *that* much code — its all just NodeJS
errata.

This commit thus opts for this approach.

== Design Notes
=== Merge Commit

With the switch from squash to merge commit, the git diff will work
propertly catching all files in the PR (rather than the last)