Delete linear history requirement (#13)
* Delete linear history requirement

This commit removes the linear history requirement on the repository,
which disallowed merges in the mainline.

* Allow multiple merging strategies

Currently there is a problem in which the "with merge commit" merge
strategy doesn't work, as the branch requires "linear" history. This
commit allows that, but also the squash commit in the case it is just
asingle commit that can be fast-forwarded onto the main branch.