Deploy, continuously, the x40.link app (#15)
Editorial Note: There was a surprising amount of work in this commit. It
                will explain it all.

Currently, the application is not deployed anywhere. This commit deploys
it.

== Design Notes
=== Google Cloud Run

The choice of runtime environment for this application is "google cloud
run". This is a container based, serverless platform that JustWorks™ and
should keep costs for this service as low as possible.

Unfortunately, using this service means working with quite a few
additional Googel Cloud primitives, but they're similar to (and not
inherently more complex than) the defacto Kubernetes versions.

=== Extending the main pipeline

Rather than create a new pipeline, authenticate it and otherwise install
tooling to deploy the application, this commit simply rewrites the
existing build pipeline to be a "build and deploy" pipeline. The
additional content is only one task invocation, and Google Cloud is
already sufficiently authenticated to work with the pipeline (at least,
from GitHub's perspective).

=== Listen Address

Google Cloud Run defaults to have the port exposed at 8080. Given this,
the current stub server implementation also defaults at 8080.

=== Additional Permisison

In order to deploy the application, Google Cloud Run requires that the
deploying service account can "assume" the role of the service account
attached to the workload, so that that account can "attach itself".

This is presently the default compute account, however seems to be
likely to change in future as we connect to additional Google services.

See:
1. https://github.com/google-github-actions/deploy-cloudrun?tab=readme-ov-file#authorization

=== Podman Flags

Reportedly, Google Cloud Run is finnicky about the manifests it runs.
The rumour is it runs on Borg and not Kubernetes, so the compliance here
might not be 100% (though it still interprets the knative definitions).

The flags on podman helped me debug this, and appear (but aren't
guaranteed) to be necessary.

See:
1. https://stackoverflow.com/a/62251882

=== infra → tofu

Because there are additional infrastructure definitions now (the cloud
run versions and the tofu versions), this commit renames the existing
infra task tofu, and the new task cloudrun. These two tasks operate on
separate lifecycles, and this resolves that ambiguity.