Make the existing links for h4n.me work
In parallel to this project work, there's the deprecation of the
h4n.link domain and replacement with h4n.me. This commit is a bit of a
hack — introducing configuration into the containerfile directly — but
it is Friday, and I want to get this done before the weekend.

Future work should remove this content in favour of a firebase datastore
backend.