Introduce a flag to customize the listen address
Currently there is a problem when running the application across
different environments: The default port requirements are different.
For example,

  On Google Cloud: The default port should be 8080 (and bound to all
                   ports) so that Google Cloud Run will send the
                   traffic to the place it is expected
  Locally:         The default port should be 80, so the browser will
                   send traffic to the default HTTP port

This commit reconciles these differences by allowing the port to be
specified when the application is invoked via the "--listen-address"
flag.

== Design Notes
=== Default Value

The default value for this application has also been altered to listen
on localhost, as this is anticipated to be the most frequent way the
application is invoked (by third parties, or people running it on
localhost).

This means the default value will no longer work in either the container
nor on Google Cloud. This means they're also adjusted
(in the Containerfile and in the knative specification)

=== No tests

The code that invokes the server is fragile, and not written in a way that
can easily be tested. Fortunately, that code was written largely to get
a proof of concept up and running, and will be refactored. In future,
this will be tested with the rest of the server package.

Given this, for now the tests are skipped.