Document the release lifecycle (#24)
This commit documents the release lifecycle. This should allow more
people to understand how to get code into production, as well
demonstrate that the author can write contribution guides to this
project.

== Design Notes
=== CONTRIBUTING.md

GitHub supports specifying a contribution guideline via the
CONTRIBUTING.md file. However, this file is less likely to be useful in
and of itself — it should be a pointer to better documentation on the
website.

See:
  * https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors