Fix the path detection (#25)
Currently, the release process does not detect "sup-paths" of the
documentation changes. For example,

  docs/foo/bar

Where bar changed. This commit seeks to address that by using the double
glob, which should expand to *all* subpaths (rather than just the
immediate subdirectories)

== Changes to docs

This commit also includes changes to docs which are largely just to
trigger the build (e.g. whitespace removal, adding a couple of
extensions to make the docs prettier)